% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_sampling_binom.R, R/bmult_sampling_mult.R
\name{binomEncompassingSampling}
\alias{binomEncompassingSampling}
\title{Get Samples From Encompassing Beta Densities}
\usage{
binomEncompassingSampling(
  inequalities,
  index = 1,
  niter = 1e+05,
  prior = FALSE,
  seed = NULL
)

binomEncompassingSampling(
  inequalities,
  index = 1,
  niter = 1e+05,
  prior = FALSE,
  seed = NULL
)
}
\arguments{
\item{inequalities}{list that contains inequality constraints for each independent inequality constrained hypotheses. 
The list is created in the generateRestrictionList function. In consists of the following elements:}

\item{index}{integer that matches for which inequality constrained hypothesis samples should be drawn. Only relevant 
when multiple independent inequality constrained hypotheses are specified. Default is 1.}

\item{niter}{number of samples.}

\item{prior}{logical. If TRUE the function will ignore the data values and sample from the prior distribution instead.}

\item{seed}{set the seed for version control.}
}
\value{
matrix of dimension niter * nsamples containing prior or posterior samples from encompassing beta distribution.

matrix of dimension niter * nsamples containing prior or posterior samples from encompassing beta distribution.
}
\description{
Get Samples From Encompassing Beta Densities

Get Samples From Encompassing Beta Densities
}
