% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_journals.R
\docType{data}
\name{journals}
\alias{journals}
\title{Prevalence of Statistical Reporting Errors}
\format{
A \code{data.frame} with 8 rows and 14 variables:
\describe{
  \item{journal}{The journal name a research article was published in.}
  \item{articles_downloaded}{The number of articles downloaded per journal}
  \item{articles_with_NHST}{The number of articles with NHST results}
  \item{perc_articles_with_NHST}{The percentage of all downloaded articles that had NHST results}
  \item{nr_NHST}{The total number of NHST results}
  \item{mean_nr_NHST_per_article_with_NHST}{The mean number of NHST results per article that had at least one NHST result}
  \item{mean_nr_NHST_per_article_all_included}{The mean number of NHST results in all downloaded articles}
  \item{errors}{The total number of errors}
  \item{dec_errors}{The total number of decision errors (i.e., an error that may have changed 
the statistical conclusion of the result).}
  \item{perc_errors}{The percentage of all results that was an error}
  \item{perc_dec_errors}{The percentage of all results that was a decision error}
  \item{perc_articles_with_errors}{The percentage of all articles that had at least one error}
  \item{perc_articles_with_dec_errors}{The percentage of all articles that had at least one error}
  \item{APAfactor}{APA factor: number of detected NHST results / total number of detected p values}
}
}
\usage{
data(journals)
}
\description{
This data set, "journals" provides a summary of statistical reporting errors (i.e., inconsistencies between reported test statistic
and reported p-value) of 16,695 research articles reporting results 
from null hypothesis significance testing (NHST). The selected articles were published in 
eight major journals in psychology between 1985 to 2013: \textit{Developmental Psychology} (DP), the \textit{Frontiers in Psychology} (FP), the 
\textit{Journal of Applied Psychology} (JAP), the \textit{Journal of Consulting and Clinical Psychology} (JCCP), 
\textit{Journal of Experimental Psychology: General} (JEPG), the \textit{Journal of Personality and Social Psychology} (JPSP), 
the \textit{Public Library of Science} (PLoS), \textit{Psychological Science} (PS).
}
\details{
In total, the authors recomputed 258,105 p-values with the software package \texttt{statcheck} 
which extracts statistics from articles and recomputes the p-values \cite{epskamp2014statcheck}. 
The anonymized dataset and the data documentation was openly available on the Open Science Framwework (\url{https://osf.io/d3ukb/}; 
\url{https://osf.io/c6ap2/}).
}
\examples{
data(journals)
# Prior specification 
# We assign a uniform Beta distribution on each binomial probability
a <- rep(1, 8)  
b <- rep(1, 8)  

counts <- journals$errors 
total  <- journals$nr_NHST
factor_levels <- levels(journals$journal)

# restricted hypothesis
Hr1 <- c('JAP , PS , JCCP , PLOS , DP , FP , JEPG < JPSP')
Hr2 <- c('JCCP < DP < JPSP')
Hr3 <- c('JAP < PS < JCCP < PLOS < DP < FP < JEPG < JPSP')

out <- binomBayesInformed(factor_levels, Hr1, a=a, b=b, counts=counts, total=total, niter = 5e4, bf_type = 'LogBFer'); out
out <- binomBayesInformed(factor_levels, Hr2, a=a, b=b, counts=counts, total=total, niter = 5e4, bf_type = 'LogBFer'); out
out <- binomBayesInformed(factor_levels, Hr3, a=a, b=b, counts=counts, total=total, niter = 5e4, bf_type = 'LogBFer'); out

summary(out)
}
\references{
Nuijten, M. B., Hartgerink, C. H., van Assen, M. A., Epskamp, S., & Wicherts, J. M.  (2016).
The prevalence of statistical reporting errors in psychology (1985–2013). Behavior Research Methods, 48, 
1205–1226.
}
\keyword{datasets}
