% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_density_tbinom.R
\name{logUnnormalizedTBinom}
\alias{logUnnormalizedTBinom}
\title{Evaluates Log Density Of Unnormalized And Unrestricted Beta Distribution}
\usage{
logUnnormalizedTBinom(xi_mat, boundaries, binom_equal, hyp_direction, a, b)
}
\arguments{
\item{xi_mat}{matrix with samples from truncated Beta density. These samples should be transformed, so they range
over the entire real line}

\item{boundaries}{list containing indices for upper and lower truncation boundaries}

\item{hyp_direction}{specifies whether the imposed inequality constrained imposes an increasing (i.e., 'smaller') or
decreasing (i.e., 'larger') trend}

\item{mult_equal}{multiplicative elements for each lower and upper bound of each inequality constrained parameter.}

\item{nr_mult_equal}{vector of multiplicative elements of collapsed parameters}

\item{nr_mult_free}{vector of multiplicative elements of free parameters}

\item{prior_and_data}{a numeric vector containing the concentration parameters (when evaluating the prior distribution)
or the updated concentration parameters (when evaluating the posterior distribution)}
}
\value{
matrix containing prior or posterior samples from truncated Dirichlet distribution
}
\description{
This function evaluates the log density of the unnormalized and unrestricted Beta distribution.
It takes probit-transformed samples from a truncated Beta distribution, transforms them back into
Beta parameter values and evaluates the density
}
