% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_density_tdir.R
\name{logUnnormalizedTDir}
\alias{logUnnormalizedTDir}
\title{Evaluates Log Density Of Unnormalized And Unrestricted Dirichlet Distribution}
\usage{
logUnnormalizedTDir(
  xi_mat,
  boundaries,
  mult_equal,
  nr_mult_equal,
  nr_mult_free,
  hyp_direction,
  prior_and_data
)
}
\arguments{
\item{xi_mat}{matrix with samples from truncated Dirichlet density. These samples should be transformed, so they range
over the entire real line}

\item{boundaries}{list containing indices for upper and lower truncation boundaries}

\item{mult_equal}{multiplicative elements for each lower and upper bound of each inequality constrained parameter.}

\item{nr_mult_equal}{vector of multiplicative elements of collapsed parameters}

\item{nr_mult_free}{vector of multiplicative elements of free parameters}

\item{hyp_direction}{specifies whether the imposed inequality constrained imposes an increasing (i.e., 'smaller') or
decreasing (i.e., 'larger') trend}

\item{prior_and_data}{a numeric vector containing the concentration parameters (when evaluating the prior distribution)
or the updated concentration parameters (when evaluating the posterior distribution)}
}
\value{
matrix containing prior or posterior samples from truncated Dirichlet distribution
}
\description{
This function evaluates the log density of the unnormalized and unrestricted Dirichlet distribution.
It takes probit-transformed samples from a truncated Dirichlet distribution, transforms them back into
Dirichlet parameter values and evaluates the density
}
