% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_equalities.R
\name{multBayesBfEquality}
\alias{multBayesBfEquality}
\title{Computes Bayes Factors For Equality Constrained Parameters}
\usage{
multBayesBfEquality(
  alphas,
  counts,
  thetas = rep(1/length(alphas), length(alphas)),
  ...
)
}
\arguments{
\item{alphas}{vector with concentration parameters}

\item{counts}{vector with data}

\item{thetas}{values of interest. Default is 1/K}

\item{...}{additional arguments (currently ignored).}
}
\value{
list consisting of the following elements:
        (1) Bf: dataframe containing the Bayes factor logBFe0, BFe0,
        (3) expected: vector with expected values
}
\description{
Computes Bayes factor for equality constrained parameters using the standard Bayesian multinomial test.
Null hypothesis H0 states that category proportions are exactly equal to predefined values.
Alternative hypothesis He states that category proportions are free to vary
}
