% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_multBayesInformed.R
\name{multBayesInformed}
\alias{multBayesInformed}
\title{Runs full Analysis}
\usage{
multBayesInformed(
  factor_levels,
  Hr,
  a,
  counts = NULL,
  niter = 50000,
  bf_type = "LogBFer",
  seed = NULL,
  ...
)
}
\arguments{
\item{factor_levels}{matrix of dimension (nsamples x nparams) with samples from truncated Dirichlet density}

\item{Hr}{character vector encoding the user specified order restriction}

\item{a}{numeric vector with concentration parameters}

\item{counts}{numeric vector with data}

\item{niter}{numeric vector with number of samples to be drawn}

\item{bf_type}{'bf_type'. The Bayes factor type can be either 'LogBFer', 'BFer', or 'BFre'. Default is 'LogBFer'}

\item{seed}{set the seed for version control.}

\item{...}{additional arguments (currently ignored).}
}
\value{
list consisting of the following elements:
        (1) BF: Bayes factor for restricted hypothesis compared to the encompassing hypothesis
        (2) restrictions: full restriction list
        (3) logBFe_equalities: log Bayes factor for equality constrained parameters
        (4) logBFe_inequalities: log Bayes factor for inequality constrained parameters
}
\description{
Runs full Analysis
}
