% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_S3methods.R
\name{print.bmult}
\alias{print.bmult}
\title{print method for class \code{bmult}}
\usage{
\method{print}{bmult}(x, ...)
}
\arguments{
\item{x}{object of class \code{bmult} as returned from \code{multBayesInformed}}

\item{...}{further arguments, currently ignored}
}
\value{
The print methods print the results and return nothing
}
\description{
print method for class \code{bmult}
}
