% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_S3methods.R
\name{samples.bmult}
\alias{samples.bmult}
\title{Extracts prior and posterior samples (if applicable) from an object of class \code{bmult}}
\usage{
\method{samples}{bmult}(x, ...)
}
\arguments{
\item{x}{object of class \code{bmult_bridge} as returned from \code{multBayesInformed}}

\item{...}{further arguments, currently ignored}
}
\value{
Returns \code{list} with prior and posterior samples (if applicable) from an object of class \code{bmult}
}
\description{
Extracts prior and posterior samples (if applicable) from an object of class \code{bmult}
}
